pyproject.toml
[project]
name = "financial-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.94.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
